Training management system
Our Training Management System (TMS) is a module of ATUM-LMS which brings efficiency into instructor-led training/class. It supports management of classroom training, uploading of trainer’s data, class attendance and resource management. It is a very useful tool allowing trainers to fix training sessions for their employees based on the need of the organization.